fix(scrollbar): move ScrollbarPadding to src/types so index.ts export resolves
