+++ src/types.ts
@@ -374,6 +374,18 @@ export interface ScrollbarOptions {
   showOnViewportEnter?: boolean;
 }
 
+/**
+ * Per-side padding for the scrollbar track.
+ * Each side defaults to the global `PADDING` constant when omitted.
+ * A plain number is shorthand for all four sides.
+ */
+export type ScrollbarPadding = number | {
+  top?: number;
+  right?: number;
+  bottom?: number;
+  left?: number;
+};
+
 // =============================================================================
 // Scrollbar (legacy — use ScrollConfig.scrollbar instead)
 // =============================================================================
